+++ client-v2/src/test/java/com/clickhouse/client/metrics/MetricsTest.java
@@ -41,7 +41,7 @@ public class MetricsTest extends BaseIntegrationTest {
         meterRegistry.clear();
         Metrics.globalRegistry.clear();
     }
-    
+
     @Test(groups = {"integration"}, enabled = true)
     public void testRegisterMetrics() throws Exception {
         ClickHouseNode node = getServer(ClickHouseProtocol.HTTP);