+++ jdbc-v2/src/test/java/com/clickhouse/jdbc/StatementTest.java
@@ -37,14 +37,7 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertNotNull;
-import static org.testng.Assert.assertNull;
-import static org.testng.Assert.assertSame;
-import static org.testng.Assert.assertThrows;
-import static org.testng.Assert.assertTrue;
-import static org.testng.Assert.fail;
+import static org.testng.Assert.*;
 
 
 @Test(groups = {"integration"})
@@ -289,7 +282,7 @@ public class StatementTest extends JdbcIntegrationTest {
     @Test(groups = {"integration"}, dataProvider = "asyncInsertSettingsDP")
     public void testInsertWithAsyncInsert(String asyncInsert, String waitAsyncInsert, int expectedUpdateCount, int expectedSelectCount, boolean fails) throws Exception {
         String tableName = "test_async_insert_param_" + asyncInsert + "_" + waitAsyncInsert + "_" + UUID.randomUUID().toString().replace("-", "_");
-        
+
         Properties props = new Properties();
         props.setProperty(ClientConfigProperties.serverSetting(ServerSettings.ASYNC_INSERT), asyncInsert);
         props.setProperty(ClientConfigProperties.serverSetting(ServerSettings.WAIT_ASYNC_INSERT), waitAsyncInsert);
@@ -466,14 +459,30 @@ public class StatementTest extends JdbcIntegrationTest {
 
     @Test(groups = {"integration"})
     public void testExecuteQueryTimeout() throws Exception {
-        try (Connection conn = getJdbcConnection()) {
+        Properties config = new Properties();
+        config.setProperty(ClientConfigProperties.ASYNC_OPERATIONS.getKey(), "true");
+        try (Connection conn = getJdbcConnection(config)) {
             try (Statement stmt = conn.createStatement()) {
-                stmt.setQueryTimeout(1);
-                assertThrows(SQLException.class, () -> {
-                    try (ResultSet rs = stmt.executeQuery("SELECT sleep(5)")) {
-                        assertFalse(rs.next());
+                long woTimeoutStart = System.currentTimeMillis();
+                final String query = "SELECT count(), sum(sipHash64(number)) " +
+                        "FROM numbers(1000000000) " +
+                        "SETTINGS max_threads = 1;";
+                try (ResultSet rs = stmt.executeQuery(query)) {
+                    assertTrue(rs.next());
+                }
+                long woTimeoutTime = System.currentTimeMillis() - woTimeoutStart;
+
+                int queryTimeoutMs = (int) (woTimeoutTime * 0.75);
+                stmt.setQueryTimeout((int) TimeUnit.MILLISECONDS.toSeconds(queryTimeoutMs));
+
+                long wTimeoutStart = System.currentTimeMillis();
+                SQLException ex = expectThrows(SQLException.class, () -> {
+                    try (ResultSet rs = stmt.executeQuery(query)) {
+                        assertTrue(rs.next());
                     }
                 });
+                long wTimeoutTime = System.currentTimeMillis() - wTimeoutStart;
+                assertTrue(Math.abs(wTimeoutTime - queryTimeoutMs) < 1000);
             }
         }
     }